Introduction

UnROOT.jl

UnROOT.Preamble

UnROOT.Streamers

UnROOT.array

UnROOT.endcheck

UnROOT.readbytes!

UnROOT.readobjany!

UnROOT.skiptobj

UnROOT.topological_sort

UnROOT.@stack

Yep.

UnROOT.array — Method.

function array(f::ROOTFile, path)

Reads an array from a branch. Currently hardcoded to Int32

source

UnROOT.Preamble — Method.

Reads the preamble of an object.

The cursor will be put into the right place depending on the data.

source

UnROOT.Streamers — Method.

function Streamers(io)

Reads all the streamers from the ROOT source.

source

UnROOT.endcheck — Method.

function endcheck(io, preamble::Preamble)

Checks if everything went well after parsing a TOBject. Used in conjuction with Preamble.

source

UnROOT.readbytes! — Method.

function readbytes!(io, b, offset, nr)

Efficient read of bytes into an existing array at a given offset

source

UnROOT.readobjany! — Method.

function readobjany!(io, tkey::TKey, refs)

The main entrypoint where streamers are parsed and cached for later use. The refs dictionary holds the streamers or parsed data which are reused when already available.

source

UnROOT.skiptobj — Method.

function skiptobj(io)

Skips a TOBject.

source

UnROOT.topological_sort — Method.

function topological_sort(streamer_infos)

Sort the streamers with respect to their dependencies and keep only those which are not defined already.

The implementation is based on https://stackoverflow.com/a/11564769/1623645

source

UnROOT.@stack — Macro.

macro stack(into, structs...)

Stack the fields of multiple structs and create a new one. The first argument is the name of the new struct followed by the ones to be stacked. Parametric types are not supported and the fieldnames needs to be unique.

Example:

@stack Baz Foo Bar

Creates Baz with the concatenated fields of Foo and Bar

source