UnROOT.Cursor — Type

The Cursor type is embeded into Branches of a TTree such that when we need to read the content of a Branch, we don't need to go through the Directory and find the TKey and then seek to where the Branch is.

Note

The io inside a Cursor is in fact only a buffer, it is NOT a io that refers to the whole file's stream.

UnROOT.Preamble — Method

Reads the preamble of an object.

The cursor will be put into the right place depending on the data.

UnROOT.ROOTFile — Method

ROOTFile(filename::AbstractString; customstructs = Dict("TLorentzVector" => LorentzVector{Float64}))

ROOTFile's constructor from a file. The customstructs dictionary can be used to pass user-defined struct as value and its corresponding fClassName (in Branch) as key such that UnROOT will know to intepret them, see interped_data.

See also: LazyTree, LazyBranch

Example

julia> f = ROOTFile("test/samples/NanoAODv5_sample.root")
ROOTFile with 2 entries and 21 streamers.
test/samples/NanoAODv5_sample.root
└─ Events
   ├─ "run"
   ├─ "luminosityBlock"
   ├─ "event"
   ├─ "HTXS_Higgs_pt"
   ├─ "HTXS_Higgs_y"
   └─ "⋮"

UnROOT.Streamers — Method

function Streamers(io)

Reads all the streamers from the ROOT source.

Base.getindex — Method

Base.getindex(ba::LazyBranch{T, J}, idx::Integer) where {T, J}

Get the idx-th element of a LazyBranch, starting at 1. If idx is within the range of ba.buffer_range, it will directly return from ba.buffer. If not within buffer, it will fetch the correct basket by calling basketarray and update buffer and buffer range accordingly.

Warning

Because currently we only cache a single basket inside LazyBranch at any given moment, access a LazyBranch from different threads at the same time can cause performance issue and incorrect event result.

UnROOT.array — Method

array(f::ROOTFile, path; raw=false)

Reads an array from a branch. Set raw=true to return raw data and correct offsets.

UnROOT.arrays — Method

arrays(f::ROOTFile, treename)

Reads all branches from a tree.

UnROOT.auto_T_JaggT — Method

auto_T_JaggT(f::ROOTFile, branch; customstructs::Dict{String, Type})

Given a file and branch, automatically return (eltype, Jaggtype). This function is aware of custom structs that are carried with the parent ROOTFile.

This is also where you may want to "redirect" classname -> Julia struct name, for example "TLorentzVector" => LorentzVector here and you can focus on LorentzVectors.LorentzVector methods from here on.

See also: ROOTFile, interped_data

UnROOT.basketarray — Method

basketarray(f::ROOTFile, path::AbstractString, ith)
basketarray(f::ROOTFile, branch::Union{TBranch, TBranchElement}, ith)

Reads actual data from ith basket of a branch. This function first calls readbasket to obtain raw bytes and offsets of a basket, then calls auto_T_JaggT followed by interped_data to translate raw bytes into actual data.

UnROOT.compressed_datastream — Method

compressed_datastream(io, tkey)

Extract all [compressionheader][rawbytes] from a TKey. This is an isolated function because we want to compartmentalize disk I/O as much as possible.

See also: decompress_datastreambytes

UnROOT.decompress_datastreambytes — Method

decompress_datastreambytes(compbytes, tkey)

Process the compressed bytes compbytes which was read out by compressed_datastream and pointed to from tkey. This function simply return uncompressed bytes according to the compression algorithm detected (or the lack of).

UnROOT.endcheck — Method

function endcheck(io, preamble::Preamble)

Checks if everything went well after parsing a TOBject. Used in conjuction with Preamble.

UnROOT.interped_data — Method

interped_data(rawdata, rawoffsets, ::Type{Vector{LorentzVector{Float64}}}, ::Type{Offsetjagg})

The interped_data method specialized for LorentzVector. This method will get called by basketarray instead of the default method for TLorentzVector branch.

UnROOT.interped_data — Method

interped_data(rawdata, rawoffsets, ::Type{T}, ::Type{J}) where {T, J<:JaggType}

The function thats interpret raw bytes (from a basket) into corresponding Julia data, based on type T and jagg type J.

In order to retrieve data from custom branches, user should defined more speialized method of this function with specific T and J. See TLorentzVector example.

UnROOT.parseTH — Method

parseTH(th::Dict{Symbol, Any})

Parse the output of TH into a tuple of counts, edges, and sumw2. A StatsBase.Histogram can then be constructed with Histogram(edges, counts). TH1 and TH2 inputs are supported.

UnROOT.readbasket — Method

readbasket(f::ROOTFile, branch, ith)
readbasketseek(f::ROOTFile, branch::Union{TBranch, TBranchElement}, seek_pos::Int)

The fundamental building block of reading read data from a .root file. Read read one basket's raw bytes and offsets at a time. These raw bytes and offsets then (potentially) get processed by interped_data.

See also: auto_T_JaggT, basketarray

UnROOT.readobjany! — Method

function readobjany!(io, tkey::TKey, refs)

The main entrypoint where streamers are parsed and cached for later use. The refs dictionary holds the streamers or parsed data which are reused when already available.

UnROOT.skiptobj — Method

function skiptobj(io)

Skips a TOBject.

UnROOT.splitup — Method

splitup(data::Vector{UInt8}, offsets, T::Type; skipbytes=0)

Given the offsets and data return by array(...; raw = true), reconstructed the actual array (with custome struct, can be jagged as well).

UnROOT.topological_sort — Method

function topological_sort(streamer_infos)

Sort the streamers with respect to their dependencies and keep only those which are not defined already.

The implementation is based on https://stackoverflow.com/a/11564769/1623645

UnROOT.unpack — Method

unpack(x::CompressionHeader)

Return the following information:

Name of compression algorithm

Level of the compression

compressedbytes and uncompressedbytes according to uproot3

UnROOT.@stack — Macro

macro stack(into, structs...)

Stack the fields of multiple structs and create a new one. The first argument is the name of the new struct followed by the ones to be stacked. Parametric types are not supported and the fieldnames needs to be unique.

Example:

@stack Baz Foo Bar

Creates Baz with the concatenated fields of Foo and Bar